Hack to manage in a different way psource and fsource
From the beginning, I uses psource and fsource in a different way.
Moreover, I have added a way to link them
automaticaly to linked notes.

In these fields, I have a list of contributors abbreviation (e.g.: for
Anselme and Larousse => A, L)
And I have linked notes in :
Sources:L (for public access)
or
p:Sources:L (for private access)

Link is colored :
- default if exists in Sources:
- orange if exists in p:Sources:
- red if not exists

For unauthorized users, text is displayed with link

TODO: changes in templates are not commited
hd/etc/module/sources.txt

TODO: don't print the link when cgl=on