Interfacing with cartography tools
1) Ascendant or descendant cartography
Add possibility to display ascendant or descendant cartography (using
external tools)
- New templates anccarto.txt and descarto.txt to prepare data as a JSON
table
- Using Leaflet for cartography displaying
- Computation is done on client side with javascript (check2.js)
- A cache for geocoding is provided as json for performances
(geoloc2.js)
- javascript location is defined in .gwf file (var_cartol)

Sample for ascendant :
http://lledieu.org/?m=A&p=stanislas&n=gamot&oc=5&v=10&t=CARTO

Sample for descendant :
http://lledieu.org/?m=D&p=stanislas&n=gamot&oc=5&v=5&t=CARTO

check2.js is available on my site : http://lledieu.org/carto2/check2.js

2) In wizard mode, check automaticaly if places are normalized. If so,
add a green check next de place. Is not, add a red cross.
Place is clickable to check2.html page (interfaced with Geonames). If
the place is not normalize, the page should help to normalize it. If the
place is normalized, page is auto redirected to google map to focus on
the place

Added :
%apply;check_place(xx)
%apply;long_married_check(xx) (long_married using check_place)

http://lledieu.org/carto2/check2.html is available on my site (it a
backend to check2.js).

Interconnection with Geonames is uning a php script. More over, I am
using a mysql database to cache results to limit request to JSON search
webservice

Naming convention for places :
- "?, <location>" means maybe
- "<location>, ?" means I can't locate it
- "k3, k2, k1" : k1 is the country
- "[k4], k3, k2, k1" : [] means no geocoding for this part

TODO:
- translation is limited
- new templates are copied from v6 (it works but could be better)