Security is back
Using Yojson to parse JSON for safe_html and sting_with_macros where it
is needed.

Face detection is now optional. To use it, set in gwf file :
face_detection=on

Edit button was broken in A2 commit.

dataTables.bootstrap4.min.js is commented to remove an erreur on
console.