Do not decide output format in views
The views `NotesDisplay.print` and `NotesDisplay.print_mod` can
output HTML or JSON content on the socket. If the user asks for JSON
result, these printers outputs a HTTP header with the appropriate MIME
format. This design breaks the separation between the dispatcher in
`gwd/request.ml` and views. In other words, a single request should
always output in the same format.

Changes in this PR:
1. Use different views according to the output format.
2. Change the output format in `request` itself.
3. Refactor a bit the code of views in `NotesDisplay`.